[HIS] Diverse Kleinarbeiten

+++ src/main/java/ch/fhnw/lems/persistence/TransportCostRepository.java
@@ -3,11 +3,11 @@ package ch.fhnw.lems.persistence;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import ch.fhnw.lems.dto.TransportCost;
+import ch.fhnw.lems.entity.TransportCost;
 
 
 //done by HiS
 @Repository
-public interface TransportCostRepository extends JpaRepository<TransportCost, Integer> {
+public interface TransportCostRepository extends JpaRepository<TransportCost, Long> {
 	
 }
